split tcpdump tests into new file
